costos , diferentes vistas 28/08/2015

+++ db/migrate/20150826163746_create_dlaborals.rb
@@ -1,11 +1,11 @@
 class CreateDlaborals < ActiveRecord::Migration
   def change
     create_table :dlaborals do |t|
-      t.integer :dias_año
-      t.integer :domingos
-      t.integer :festivos
-      t.integer :vacaciones
-      t.integer :total_dias
+      t.float :dias_año
+      t.float :domingos
+      t.float :festivos
+      t.float :vacaciones
+      t.float :total_dias
 
       t.timestamps null: false
     end
