+++ db/migrate/20150826164639_create_sueldos.rb
@@ -1,12 +1,12 @@
 class CreateSueldos < ActiveRecord::Migration
   def change
     create_table :sueldos do |t|
-      t.float :salario
-      t.float :salario_porcentaje
-      t.float :subsidio
-      t.float :subsidio_procentaje
-      t.float :total_sueldo
-      t.float :total_porcentaje
+      t.integer :salario
+      t.integer :salario_porcentaje
+      t.integer :subsidio
+      t.integer :subsidio_procentaje
+      t.integer :total_sueldo
+      t.integer :total_porcentaje
 
       t.timestamps null: false
     end
