+++ db/migrate/20150826204533_create_prestaciones.rb
@@ -1,16 +1,16 @@
 class CreatePrestaciones < ActiveRecord::Migration
   def change
     create_table :prestaciones do |t|
-      t.float :cesantias
-      t.float :cesantias_porcentaje
-      t.float :prima
-      t.float :prima_procentaje
-      t.float :vacaciones
-      t.float :vacaciones_porcentaje
-      t.float :intereses_cesantias
-      t.float :intereses_porcenjate
-      t.float :total_prestacion
-      t.float :total_porcentaje
+      t.integer :cesantias
+      t.integer :cesantias_porcentaje
+      t.integer :prima
+      t.integer :prima_procentaje
+      t.integer :vacaciones
+      t.integer :vacaciones_porcentaje
+      t.integer :intereses_cesantias
+      t.integer :intereses_porcenjate
+      t.integer :total_prestacion
+      t.integer :total_porcentaje
       t.belongs_to :sueldo, index: true, foreign_key: true
 
       t.timestamps null: false
